+++ Source/UnrealCAPIHost/Private/uec_host_gameplay_example.c
@@ -1,0 +1,2 @@
+/* Compile the documented example as a C translation unit in the Unreal host. */
+#include "../../../examples/c_gameplay/c_gameplay.c"