+++ Source/UnrealCAPIHost/UnrealCAPIHost.Build.cs
@@ -1,4 +1,5 @@
 using UnrealBuildTool;
+using System.IO;
 
 public class UnrealCAPIHost : ModuleRules
 {
@@ -7,5 +8,7 @@ public class UnrealCAPIHost : ModuleRules
         PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
         PublicDependencyModuleNames.Add("Core");
         PrivateDependencyModuleNames.AddRange(new[] { "CoreUObject", "Engine", "UnrealCAPI" });
+        PrivateIncludePaths.Add(Path.GetFullPath(
+            Path.Combine(ModuleDirectory, "../../examples/c_gameplay")));
     }
 }